Tests/test_utils.py
import unittest
from unittest.mock import patch, Mock
from pyspark.sql import DataFrame
import os

# Import the functions to be tested
from utils.utils import Converter, checkpointLocation, get_root_dir

class TestConverterFunctions(unittest.TestCase):
    @classmethod
    def setUpClass(cls):
        # Set up any necessary resources or configuration for all test cases
        pass

    @classmethod
    def tearDownClass(cls):
        # Clean up any resources or configuration after all test cases
        pass

    def setUp(self):
        # Set up any necessary resources or configuration for each test case
        self.sample_data = DataFrame()  # Create a sample DataFrame
        self.config = {"source_col1": "modified_col1", "source_col2": "modified_col2"}

    def tearDown(self):
        # Clean up any resources or configuration after each test case
        pass

    def test_converter_with_write_flag_true(self):
        # Test the Converter function when write_flag is True
        with patch('your_module.DataFrame', autospec=True) as mock_df:
            # Mock the DataFrame class
            Converter(self.sample_data, self.config, write_flag=True)
            # Add assertions here to check if the DataFrame is modified as expected

    def test_converter_with_write_flag_false(self):
        # Test the Converter function when write_flag is False
        with patch('your_module.DataFrame', autospec=True) as mock_df:
            # Mock the DataFrame class
            Converter(self.sample_data, self.config, write_flag=False)
            # Add assertions here to check if the DataFrame is modified as expected

    @patch('your_module.os.makedirs')
    def test_checkpoint_location_with_default_path(self, mock_makedirs):
        # Test the checkpointLocation function with the default path
        path = checkpointLocation()
        # Add assertions here to check if the path is constructed correctly and os.makedirs is called

    @patch('your_module.os.makedirs')
    def test_checkpoint_location_with_custom_path(self, mock_makedirs):
        # Test the checkpointLocation function with a custom path
        custom_path = "/custom/path/to/checkpoint"
        path = checkpointLocation(custom_path)
        # Add assertions here to check if the path is constructed correctly and os.makedirs is called with the custom path

    def test_get_root_dir(self):
        # Test the get_root_dir function
        with patch('your_module.os.path.exists', return_value=True):
            root_dir = get_root_dir()
            # Add assertions here to check if the correct root directory is returned
